% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_data.R
\name{available_data}
\alias{available_data}
\title{Get Available \pkg{lexicon} Data}
\usage{
available_data(regex = NULL, ...)
}
\arguments{
\item{regex}{A regex to search for within the data columns.}

\item{\ldots}{Other arguments passed to \code{grep}.}
}
\value{
Returns a data.frame
}
\description{
See available \pkg{lexicon} data a data.frame.
}
\examples{
available_data()
available_data('hash_')
available_data('hash_sentiment')
available_data('python')
available_data('prof')
available_data('English')
available_data('Stopword')
}
