% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_names.R
\docType{data}
\name{common_names}
\alias{common_names}
\title{First Names (U.S.)}
\format{A character vector with 5493 elements}
\usage{
data(common_names)
}
\description{
A dataset containing 1990 U.S. census data on first names.
}
\references{
http://www.census.gov
}
\keyword{datasets}
