% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_words.R
\docType{data}
\name{function_words}
\alias{function_words}
\title{Function Words}
\format{A character vector with 350 elements}
\usage{
data(function_words)
}
\description{
A vector of function words from
\href{http://myweb.tiscali.co.uk/wordscape/museum/funcword.html}{John and Muriel Higgins's list}
used for the text game ECLIPSE.  The list is augmented with additional
contractions from \code{\link[lexicon]{key_contractions}}.
}
\section{Copyright}{
 John Higgins
}

\references{
\url{http://myweb.tiscali.co.uk/wordscape/museum/funcword.html}
}
\keyword{datasets}
