% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grady_augmented.R
\docType{data}
\name{grady_augmented}
\alias{grady_augmented}
\title{Augmented List of Grady Ward's English Words and Mark Kantrowitz's Names List}
\format{A character vector with 122,806 elements}
\usage{
data(grady_augmented)
}
\description{
A dataset containing a vector of Grady Ward's English words augmented with
Mark Kantrowitz's names list, other proper nouns, and contractions.
}
\details{
A dataset containing a vector of Grady Ward's English words
augmented with proper nouns (U.S. States, Countries, Mark Kantrowitz's Names List,
and months) and contractions. That dataset is augmented for spell checking purposes.
}
\references{
Moby Thesaurus List by Grady Ward
}
\keyword{datasets}
