% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_emoticons.R
\docType{data}
\name{hash_emoticons}
\alias{hash_emoticons}
\title{Emoticons}
\format{A data.table with 144 rows and 2 variables}
\usage{
data(hash_emoticons)
}
\description{
A \pkg{data.table} key containing common emoticons (adapted from
Wikipedia's Page semi-protected 'List of emoticons').
}
\details{
\itemize{
  \item x. The graphic representation of the emoticon
  \item y. The meaning of the emoticon
}
}
\section{License}{
 https://creativecommons.org/licenses/by-sa/3.0/legalcode
}

\examples{
\dontrun{
library(data.table)
hash_emoticons[c(':-(', '0:)')]
}
}
\references{
https://en.wikipedia.org/wiki/List_of_emoticons
}
\keyword{datasets}
