% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_internet_slang.R
\docType{data}
\name{hash_internet_slang}
\alias{hash_internet_slang}
\title{List of Internet Slang and Corresponding Meanings}
\format{A data frame with 175 rows and 2 variables}
\usage{
data(hash_internet_slang)
}
\description{
A dataset containing Internet slang terms and corresponding meaning.  The data
set is an augmented version of \url{http://www.smart-words.org/abbreviations/text.html}.
}
\details{
\itemize{
  \item x. The slang term.
  \item y. The meaning.
}
}
\references{
Possel, H. (n.d.). English language smart words.  Retrieved from 
http://www.smart-words.org \cr \cr
\url{http://www.smart-words.org/abbreviations/text.html}
}
\keyword{datasets}
