% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_nrc_emotion.R
\docType{data}
\name{hash_nrc_emotions}
\alias{hash_nrc_emotions}
\title{NRC Emotion Table}
\format{A data frame with 8265 rows and 2 variables}
\usage{
data(hash_nrc_emotions)
}
\description{
A \pkg{data.table} dataset containing a filtered version of Mohammad &
Turney', P. D.'s (2010) emotion word list as lookup table.
}
\details{
\itemize{ 
  \item token. A search token indicating emotion.
  \item emotion. An accompanying emotion assocatiated with the token.
}
}
\references{
http://www.purl.com/net/lexicons  \cr \cr
Mohammad, S. M. & Turney, P. D. (2010) Emotions evoked by common words and
phrases: Using Mechanical Turk to create an emotion lexicon, In Proceeding of
Workshop on Computational Approaches to Analysis and Generation of Emotion in
Text, 26-34.
}
\keyword{datasets}
