% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_jockers_rinker.R
\docType{data}
\name{hash_sentiment_jockers_rinker}
\alias{hash_sentiment_jockers_rinker}
\title{Combined Jockers & Rinker Polarity Lookup Table}
\format{A data frame with 11,710 rows and 2 variables}
\usage{
data(hash_sentiment_jockers_rinker)
}
\description{
A \pkg{data.table} dataset containing a combined and augmented version of
Jockers (2017) & Rinker's augmented Hu & Liu (2004) positive/negative word
list as sentiment lookup values.
}
\details{
\itemize{
  \item x. Words
  \item y. Sentiment
}
}
\references{
Jockers, M. L. (2017). Syuzhet: Extract sentiment and plot arcs
from Text. Retrieved from https://github.com/mjockers/syuzhet \cr \cr

Hu, M., & Liu, B. (2004). Mining and summarizing customer reviews.
Proceedings of the ACM SIGKDD International Conference on Knowledge Discovery
& Data Mining (KDD-2004). Seattle, Washington.
}
\keyword{datasets}
