% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_loughran_mcdonald.R
\docType{data}
\name{hash_sentiment_loughran_mcdonald}
\alias{hash_sentiment_loughran_mcdonald}
\title{Loughran-McDonald Polarity Table}
\format{A data frame with 2,702 rows and 2 variables}
\usage{
data(hash_sentiment_loughran_mcdonald)
}
\description{
A \pkg{data.table} dataset containing an filtered version of Loughran &
McDonald's (2016) positive/negative financial word list as sentiment lookup
values.
}
\details{
\itemize{
  \item x. Words
  \item y. Sentiment values
}
}
\section{License}{
 The original authors note the data is available for 
non-commercial, research use: "The data compilations provided on 
this website are for use by individual researchers.".  For more details see:
https://sraf.nd.edu/textual-analysis/resources/#Master%20Dictionary.
}

\section{Copyright}{
 Copyright holder University of Notre Dame
}

\references{
Loughran, T. and McDonald, B. (2016). Textual analysis in
accounting and finance: A survey. Journal of Accounting Research 54(4),
1187-1230. doi: 10.2139/ssrn.2504147 \cr \cr
\url{https://sraf.nd.edu/textual-analysis/resources/#Master\%20Dictionary}
}
\keyword{datasets}
