% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_socal_google.R
\docType{data}
\name{hash_sentiment_socal_google}
\alias{hash_sentiment_socal_google}
\title{SO-CAL Google Polarity Table}
\format{A data frame with 3272 rows and 2 variables}
\usage{
data(hash_sentiment_socal_google)
}
\description{
A \pkg{data.table} dataset containing a version of Taboada, Brooke, Tofiloski,
Voll, & Stede's (2011) positive/negative word list as sentiment lookup
values.
}
\details{
\itemize{
  \item x. Words
  \item y. Sentiment values
}
}
\section{License}{
 The original license states:
This work is licensed under a Creative Commons 
Attribution-NonCommercial-ShareAlike 4.0 International License.
https://creativecommons.org/licenses/by-nc-sa/4.0/
}

\references{
Taboada, M., Brooke, J., Tofiloski, M., Voll, K., and Stede, M.
(2011). Lexicon-based methods for sentiment analysis. Computational
Linguistics, 37(2). 267-307. \cr \cr
\url{https://github.com/sfu-discourse-lab/SO-CAL}
}
\keyword{datasets}
