% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_grade.R
\docType{data}
\name{key_grade}
\alias{key_grade}
\title{Grades Data Set}
\format{A data frame with 15 rows and 2 variables}
\usage{
data(key_grade)
}
\description{
A dataset containing common grades.
}
\details{
\itemize{
  \item x. The graphic representation of the grade
  \item y. The meaning of the grade
}
}
\keyword{datasets}
