% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal_loughran_mcdonald.R
\docType{data}
\name{modal_loughran_mcdonald}
\alias{modal_loughran_mcdonald}
\title{Loughran-McDonald Modal List}
\format{A data frame with 61 rows and 2 variables}
\usage{
data(modal_loughran_mcdonald)
}
\description{
A dataset containing a character vector of Loughran & McDonald's (2016)
modal list.  Wikipedia states: "A modal verb is a type of verb that is used
to indicate modality - that is: likelihood, ability, permission and obligation."
}
\details{
\itemize{
  \item modal.
  \item strength.
}
}
\section{License}{
 The original authors note the data is available for 
non-commercial, research use: "The data compilations provided on 
this website are for use by individual researchers.".  For more details see:
https://sraf.nd.edu/textual-analysis/resources/#Master%20Dictionary.
}

\section{Copyright}{
 Copyright holder University of Notre Dame
}

\references{
Loughran, T. and McDonald, B. (2016). Textual analysis in
accounting and finance: A survey. Journal of Accounting Research 54(4),
1187-1230. doi: 10.2139/ssrn.2504147 \cr \cr
\url{https://sraf.nd.edu/textual-analysis/resources/#Master\%20Dictionary}
}
\keyword{datasets}
