% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_action_verb.R
\docType{data}
\name{pos_action_verb}
\alias{pos_action_verb}
\title{Action Word List}
\format{A character vector with 1569 elements}
\usage{
data(pos_action_verb)
}
\description{
A dataset containing a vector of action words.  This is a subset of the
Moby project: Moby Part-of-Speech.
}
\details{
From Grady Ward's Moby project:
"This second edition is a particularly thorough revision of the original Moby
Part-of-Speech. Beyond the fifteen thousand new entries, many thousand more
entries have been scrutinized for correctness and modernity. This is
unquestionably the largest P-O-S list in the world. Note that the many included
phrases means that parsing algorithms can now tokenize in units larger than a
single word, increasing both speed and accuracy."  Originally downloaded
from: http://icon.shef.ac.uk/Moby
}
\keyword{datasets}
