% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_df_pronouns.R
\docType{data}
\name{pos_df_pronouns}
\alias{pos_df_pronouns}
\title{Pronouns}
\format{A data frame with 34 rows and 5 variables}
\usage{
data(pos_df_pronouns)
}
\description{
A dataset containing pronouns categorized by type, singular, point_of_view,
and use.  Note that 'you', and 'yours' appear twice because 'you' can be
singular or plural.
}
\details{
\itemize{
  \item pronoun. The pronoun.
  \item type. The pronoun type; either "personal", "reflexive", or "possessive".
  \item singular. logical.  If \code{TRUE} the pronoun is singular, otherwise it's plural.
  \item point_of_view. The point of view; either "first", "second", or "third".
}
}
\references{
http://www.english-grammar-revolution.com/list-of-pronouns.html
}
\keyword{datasets}
