% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_interjections.R
\docType{data}
\name{pos_interjections}
\alias{pos_interjections}
\title{Interjections}
\format{A character vector with 139 elements}
\usage{
data(pos_interjections)
}
\description{
Vidar Holen's dataset containing a character vector of common interjections
compiled from: http://www.vidarholen.net/contents/interjections.
}
\references{
\url{http://www.vidarholen.net/contents/interjections/}
}
\keyword{datasets}
