% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_preposition.R
\docType{data}
\name{pos_preposition}
\alias{pos_preposition}
\title{Preposition Words}
\format{A character vector with 162 elements}
\usage{
data(pos_preposition)
}
\description{
A dataset containing a vector of common prepositions.
}
\keyword{datasets}
