% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity_alvarez.R
\docType{data}
\name{profanity_alvarez}
\alias{profanity_alvarez}
\title{Alejandro U. Alvarez's List of Profane Words}
\format{A character vector with 438 elements}
\usage{
data(profanity_alvarez)
}
\description{
A dataset containing a character vector of profane words from Alejandro U.
Alvarez.
}
\section{TermsOfUse}{
 https://archive.org/about/terms.php
}

\references{
\url{https://web.archive.org/web/20130704010355/http://urbanoalvarez.es:80/blog/2008/04/04/bad-words-list/}
}
\keyword{datasets}
