% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity_arr_bad.R
\docType{data}
\name{profanity_arr_bad}
\alias{profanity_arr_bad}
\title{Stackoverflow user2592414's List of Profane Words}
\format{A character vector with 343 elements}
\usage{
data(profanity_arr_bad)
}
\description{
A dataset containing a character vector of profane words from Stackoverflow
user2592414.
}
\section{License}{
 The Stackoverflow data is Creative Commons.  See 
https://creativecommons.org/licenses/by-sa/3.0/ for License information.
}

\references{
\url{https://stackoverflow.com/a/17706025/1000343}
}
\keyword{datasets}
