% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity_zac_anger.R
\docType{data}
\name{profanity_zac_anger}
\alias{profanity_zac_anger}
\title{Zac Anger's List of Profane Words}
\format{A character vector with 3,076 elements}
\usage{
data(profanity_zac_anger)
}
\description{
A dataset containing a character vector of profane words from Zac Anger.
}
\section{License}{
 The original authors note the data allows the following: 
"Everyone is permitted to copy and distribute verbatim or modified copies of 
this license document, and changing it is allowed as long as the name is 
changed." https://github.com/zacanger/profane-words/blob/master/LICENSE.md
}

\references{
\url{https://github.com/zacanger/profane-words}
}
\keyword{datasets}
