% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dolch.R
\docType{data}
\name{sw_dolch}
\alias{sw_dolch}
\title{Leveled Dolch List of 220 Common Words}
\format{A character vector with 220 elements}
\usage{
data(sw_dolch)
}
\description{
Edward William Dolch's list of 220 Most Commonly Used Words by reading level.
}
\details{
Dolch's Word List made up 50-75\% of all printed text in 1936.
\itemize{
  \item Word. The word
  \item Level. The reading level of the word
}
}
\references{
Dolch, E. W. (1936). A basic sight vocabulary. Elementary School
Journal, 36, 456-460.
}
\keyword{datasets}
