% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_fry_1000.R
\docType{data}
\name{sw_fry_1000}
\alias{sw_fry_1000}
\title{Fry's 1000 Most Commonly Used English Words}
\format{A character vector with 1000 elements}
\usage{
data(sw_fry_1000)
}
\description{
A stopword list containing a character vector of stopwords.
}
\details{
Fry's 1000 Word List makes up 90\% of all printed text.
}
\references{
Fry, E. B. (1997). Fry 1000 instant words. Lincolnwood, IL:
Contemporary Books.
}
\keyword{datasets}
