% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_jockers.R
\docType{data}
\name{sw_jockers}
\alias{sw_jockers}
\title{Matthew Jocker's Expanded Topic Modeling Stopword List}
\format{A character vector with 5,902 elements}
\usage{
data(sw_jockers)
}
\description{
A dataset containing a character vector of Jocker's stopwords he used for
topic modeling.  He later resorted to eliminating everything but nouns:
\url{http://www.matthewjockers.net/2013/04/12/secret-recipe-for-topic-modeling-themes/}.
}
\references{
\url{http://www.matthewjockers.net/materials/uwm-2013}
}
\keyword{datasets}
